pi_can/requirements.txt
python-can==4.2.2